ConfigureAutoLinkCommandHandler
